Update patch set 2
Patch Set 2:

I agree the suffix ‘Test’ looks weird for classes in src/main, however these are abstract utility classes meant to be used to write tests, so they must be part of src/main; other modules will depend on them (It is like adding EasyMock as a dependency). We could try renaming them, but I think they should remain in src/main. 

persistence-test-framework is a different project so applications depend on it for integration testing (independent of any implementation). It is not part of persistence-util-test because everything inside persistence-util is more general purpose, persistence-util could actually be moved outside persistence.

Any thoughts?

Patch-set: 2
Label: Verified=0